services/api/booking_integration_service/requirements_simple.txt
# Simple requirements - no ODBC driver needed!
pymssql==2.2.8
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
